Update the RingHelper.sol library to emit an event whenever there is negative waiveFee
After speaking to @Brechtpd there was no good way to know when a negative waiveFeePercentage is settled from an event's perspective, because that order's fee is emitted as 0. The code changes in this file reflect the need for an event to be emitted whenever a rebate occurs. 

This code has been tested via Ganache and Dolomite's backend application code in Scala. This is the last change we need to the protocol for launch, as everything else has been tested and is working, as expected.